+++ src/IQR/IQR.ts
@@ -1,7 +1,7 @@
 import type {BigSource} from 'big.js';
 import Big from 'big.js';
 import {BigIndicatorSeries, NumberIndicatorSeries} from '../Indicator.js';
-import {getFasterMedian, getMedian} from '../util/getMedian.js';
+import {getFasterQuartile, getQuartile} from '../util/getQuartile.js';
 
 /**
  * Interquartile Range (IQR)
@@ -17,41 +17,6 @@ export class IQR extends BigIndicatorSeries<Big> {
 
   constructor(public readonly interval: number) {
     super();
-  }
-
-  private calculateQuartile(sorted: Big[], quartile: 0.25 | 0.75): Big {
-    const n = sorted.length;
-
-    // Find the median index
-    const medianIndex = Math.floor(n / 2);
-
-    if (quartile === 0.25) {
-      // Q1 calculation
-      if (n % 2 === 0) {
-        // Even number of items: the lower half is the first n/2 items
-        return getMedian(sorted.slice(0, medianIndex));
-      }
-      // Odd number of items: the lower half excludes the median
-      return getMedian(sorted.slice(0, medianIndex));
-    } else if (quartile === 0.75) {
-      // Q3 calculation
-      if (n % 2 === 0) {
-        // Even number of items: the upper half is the last n/2 items
-        return getMedian(sorted.slice(medianIndex));
-      }
-      // Odd number of items: the upper half excludes the median
-      return getMedian(sorted.slice(medianIndex + 1));
-    }
-
-    // Fallback to standard quantile calculation for non-quartile values
-    const pos = (sorted.length - 1) * quartile;
-    const base = Math.floor(pos);
-    const rest = pos - base;
-
-    if (sorted[base + 1] !== undefined) {
-      return sorted[base].plus(sorted[base + 1].minus(sorted[base]).times(rest));
-    }
-    return sorted[base];
   }
 
   update(value: BigSource, replace: boolean): Big | null {
@@ -71,11 +36,8 @@ export class IQR extends BigIndicatorSeries<Big> {
       return null;
     }
 
-    // Sort a copy for quartile calculation
-    const sorted = [...this.values].sort((a, b) => a.cmp(b));
-
-    const q1 = this.calculateQuartile(sorted, 0.25);
-    const q3 = this.calculateQuartile(sorted, 0.75);
+    const q1 = getQuartile(this.values, 0.25);
+    const q3 = getQuartile(this.values, 0.75);
 
     return this.setResult(q3.minus(q1), replace);
   }
@@ -103,49 +65,9 @@ export class FasterIQR extends NumberIndicatorSeries<number> {
       return null;
     }
 
-    const sorted = [...this.values].sort((a, b) => a - b);
-
-    const q1 = quartileNumberWikipedia(sorted, 0.25);
-    const q3 = quartileNumberWikipedia(sorted, 0.75);
+    const q1 = getFasterQuartile(this.values, 0.25);
+    const q3 = getFasterQuartile(this.values, 0.75);
 
     return this.setResult(q3 - q1, replace);
   }
-}
-
-/**
- * Calculate a quartile using the Wikipedia method for number arrays
- */
-function quartileNumberWikipedia(sorted: number[], q: number): number {
-  const n = sorted.length;
-
-  // Find the median index
-  const medianIndex = Math.floor(n / 2);
-
-  if (q === 0.25) {
-    // Q1 calculation
-    if (n % 2 === 0) {
-      // Even number of items: the lower half is the first n/2 items
-      return getFasterMedian(sorted.slice(0, medianIndex));
-    }
-    // Odd number of items: the lower half excludes the median
-    return getFasterMedian(sorted.slice(0, medianIndex));
-  } else if (q === 0.75) {
-    // Q3 calculation
-    if (n % 2 === 0) {
-      // Even number of items: the upper half is the last n/2 items
-      return getFasterMedian(sorted.slice(medianIndex));
-    }
-    // Odd number of items: the upper half excludes the median
-    return getFasterMedian(sorted.slice(medianIndex + 1));
-  }
-
-  // Fallback to standard quantile calculation for non-quartile values
-  const pos = (sorted.length - 1) * q;
-  const base = Math.floor(pos);
-  const rest = pos - base;
-
-  if (sorted[base + 1] !== undefined) {
-    return sorted[base] + (sorted[base + 1] - sorted[base]) * rest;
-  }
-  return sorted[base];
 }
